add handler and handler_test
